-Wall police
Tweak .cabal a little, add some LANGUAGE pragmas, rename some variables, add type sigs. etc.
-Wall is now clean. A side issue: the current version installs fine under a modern installation of GHC 6.8.2,
but that's not true of the Hackage package, which seems fairly out-of-date. Any chance you could upload new packages
of this and MissingH to lessen the brokeness? A lot of other packages seem to depend on ConfigFile.